Создал интерфейс IHealth и использую float, вместо int для здоровья

+++ Assets/Scripts/Interfaces/IMedicinable.cs
@@ -2,6 +2,6 @@
 {
     public interface IMedicinable
     {
-        void Treat(int amount);
+        void Treat(float amount);
     }
 }
